altered the table to hold the latitude and longitude

+++ backend/userAuthService/src/database/databaseSchema.sql
@@ -18,5 +18,5 @@ ADD About NVARCHAR(MAX),
     ProfilePhoto NVARCHAR(255);
 
  ALTER Table Users 
- ADD Latitude INT,
-     Longitude INT;
+ ADD Latitude FLOAT,
+     Longitude FLOAT;
